requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
python-dotenv
aiosqlite
pytest
httpx
PyJWT
google-auth
requests